db databoot (#100)
* add SQL formatted structure dump to db/init

* mount the db/init directory to automatically load database schema

* move data ingestion into database docker init

* seed the ar_internal_metadata table

this avoids a database environment guard introduced in rails5,
that we're not compensating for by running db:seed

* ensure that a test database exists

this version of rails doens't seem to want to create the test database
when calling `rails db:create`.  This creates an (empty) database for
the test database.  From here, we can call rails `db:test:prepare`, etc
as generally expected for the testing workflow.