Add dependency auditing to AWBW (#76)
* Add dependency auditing to AWBW

Adds dependency auditing using GitHub's dependabot
Based on the Homeward Tails implementation here
https://github.com/rubyforgood/homeward-tails/blob/main/.github/dependabot.yml
with the interval initially set to weekly to test the workflow.

If we are comfortable with a more frequent upgrade schedule,
the configuration can be changed to daily.

* Add limit for open pull requests

* Add TODO for group updates to dependabot

Added a TODO for adding group updates to dependabot in the dependabot workflow so that it doesn't get forgotten since we are not sure how grouping will fit with the current structure of the workflows.